Prematurely fetch games of the round
Fetch all the games of the round as soon as we get an event for a game
we haven't seen before. This allows us to surface the current state of
games in the FE, and saves having to make a call to the squiggle REST
API for each new game.